refactor: 역할에 따라 supabase 관련 타입 분리
